requirements.txt

fastapi
uvicorn
python-multipart
openai
pymupdf
reportlab
openai-agents
streamlit
websockets
mcp
fastmcp
python-docx
odfpy
python-pptx
